lab7/predictionpipeline/requirements.txt
pandas==1.0.5
keras==2.4.0
numpy==1.16.0
google-cloud-storage